Refactor for 64-bit safety and code maintainability
Updated return types, replaced unsafe functions, and improved
type safety for 64-bit compatibility. Added "TODO" tags to
#pragma messages for better task tracking. Key changes include:

- Changed `LONG` to `INT_PTR` for `pGetLayerName` in multiple files.
- Replaced `_ltoa` with `_ltoa_s` for safer string conversion.
- Replaced `mkdir` with `_mkdir` for Windows compatibility.
- Standardized `GetSize()` calls with `static_cast<int>` for safety.
- Added `GetCount` methods to `CSeg` and `CSegs` for type safety.
- Updated `WndProcKeyPlanOnMouseMove` to align with Windows API.

Refactor and improve code safety and consistency

- Changed `LONG` to `INT_PTR` for `pGetLayerName` to ensure 64-bit compatibility.
- Increased warning level to `Level4` in project files for stricter code checks.
- Replaced unsafe functions (`_ltoa`, `mkdir`) with safer alternatives (`_ltoa_s`, `_mkdir`).
- Corrected file name casing (`Stdafx` to `stdafx`) for case-sensitive systems.
- Added `TODO` prefixes to `#pragma` messages for better task tracking.
- Replaced raw integer casts with `static_cast<int>` for improved type safety.
- Updated Windows API parameter types for consistency (`UINT` to `WPARAM`).
- Refactored loops and calculations for clarity and robustness.

+++ PrimCSpline.cpp
@@ -99,7 +99,7 @@ CPnt CPrimCSpline::GoToNxtCtrlPt() const
 {
 	CPnt pt;
 
-	int i = m_pts.GetSize() - 1;
+	int i = static_cast<int>(m_pts.GetSize()) - 1;
 	
 	if (mS_dRel <= DBL_EPSILON)
 		pt = m_pts[i];
